Refactored Tile Placement Structure
Began using Raycasts for detecting tile position. No longer leaving artifacts when moving mouse too fast. Also decoupled the input from the script. Now it recieves the input from the Input System other than the raycast. Added stubs for creating structures. Added the ability to toggle the preview tile as a proof of concept. It is bound to the C key.